Big refactor to remove /dashboard from all URLs
