+++ cypress/integration/post-foraging.spec.js
@@ -4,11 +4,11 @@ describe('Checks My Feed works as expected', () => {
     cy.setLocalStorage('welcome-modal-july-2020', 'seen')
   })
   it('User visits the My Feed page', () => {
-    cy.visit('/dashboard/my-feed')
+    cy.visit('/my-feed')
     cy.get('[data-testid=my-feed-header]').should('exist')
   })
   it('User views the first post in the feed', () => {
-    cy.visit('/dashboard/my-feed')
+    cy.visit('/my-feed')
     cy.get('[data-testid=my-feed-post-card]')
       .first()
       .then(($post) => {
@@ -18,7 +18,7 @@ describe('Checks My Feed works as expected', () => {
       })
   })
   it('User views their "my posts" page', () => {
-    cy.visit('/dashboard/my-feed')
+    cy.visit('/my-feed')
     cy.get('[data-testid=my-posts-nav-link]').click()
     cy.get('[data-testid=my-posts-header]')
   })
