+++ cypress/integration/update-account-info.spec.js
@@ -2,7 +2,7 @@ describe('Update account information', () => {
   beforeEach(() => {
     cy.login()
     cy.setLocalStorage('welcome-modal-july-2020', 'seen')
-    cy.visit('/dashboard/settings/profile')
+    cy.visit('/settings/profile')
   })
   it('User updates handle', () => {
     // TODO: Write test
