+++ packages/web/pages/dashboard/settings/account.tsx
@@ -1,32 +1,14 @@
-import React from 'react'
-import { NextPage } from 'next'
+import { useRouter } from 'next/router'
 import { withApollo } from '@/lib/apollo'
-import SettingsPageLayout from '@/components/Layouts/SettingsPageLayout'
-import AuthGate from '@/components/AuthGate'
-import UpdatePasswordForm from '@/components/Dashboard/Settings/UpdatePasswordForm'
 
-const Account: NextPage = () => {
-  return (
-    <AuthGate>
-      <>
-        <SettingsPageLayout>
-          <div className="forms-container">
-            <UpdatePasswordForm />
-          </div>
-        </SettingsPageLayout>
-        <style jsx>{`
-          .forms-container {
-            width: 100%;
-            max-width: 1008px;
-          }
-        `}</style>
-      </>
-    </AuthGate>
-  )
+const OldAccountPage = () => {
+  const router = useRouter()
+  if (typeof window !== 'undefined') {
+    router.push({
+      pathname: `/settings/account`,
+    })
+  }
+  return null
 }
 
-Account.getInitialProps = async () => ({
-  namespacesRequired: ['common', 'settings'],
-})
-
-export default withApollo(Account)
+export default withApollo(OldAccountPage)
