+++ packages/web/pages/dashboard/settings/subscription.tsx
@@ -1,43 +1,14 @@
-import React from 'react'
-import { NextPage } from 'next'
+import { useRouter } from 'next/router'
 import { withApollo } from '@/lib/apollo'
-import SettingsPageLayout from '@/components/Layouts/SettingsPageLayout'
-import AuthGate from '@/components/AuthGate'
-import SubscriptionForm from '@/components/Dashboard/Settings/Subscription/SubscriptionForm'
-import { useSubscriptionSettingsPageQuery } from '@/generated/graphql'
-import theme from '@/theme'
 
-const Subscription: NextPage = () => {
-  const { loading, data, refetch: refetchUser } = useSubscriptionSettingsPageQuery()
-
-  return (
-    <AuthGate>
-      <>
-        <SettingsPageLayout>
-          {data?.currentUser && !loading && (
-            <div className="forms-container">
-              <SubscriptionForm user={data.currentUser} onSuccess={refetchUser} />
-            </div>
-          )}
-        </SettingsPageLayout>
-        <style jsx>{`
-          .forms-container {
-            width: 100%;
-            max-width: 1008px;
-          }
-          h1 {
-            ${theme.typography.headingLG}
-            text-align: center;
-            margin-bottom: 20px;
-          }
-        `}</style>
-      </>
-    </AuthGate>
-  )
+const OldSubscriptionPage = () => {
+  const router = useRouter()
+  if (typeof window !== 'undefined') {
+    router.push({
+      pathname: `/my-feed`,
+    })
+  }
+  return null
 }
 
-Subscription.getInitialProps = async () => ({
-  namespacesRequired: ['common', 'settings'],
-})
-
-export default withApollo(Subscription)
+export default withApollo(OldSubscriptionPage)
