+++ packages/web/pages/settings/subscription.tsx
@@ -1,0 +1,43 @@
+import React from 'react'
+import { NextPage } from 'next'
+import { withApollo } from '@/lib/apollo'
+import SettingsPageLayout from '@/components/Layouts/SettingsPageLayout'
+import AuthGate from '@/components/AuthGate'
+import SubscriptionForm from '@/components/Dashboard/Settings/Subscription/SubscriptionForm'
+import { useSubscriptionSettingsPageQuery } from '@/generated/graphql'
+import theme from '@/theme'
+
+const Subscription: NextPage = () => {
+  const { loading, data, refetch: refetchUser } = useSubscriptionSettingsPageQuery()
+
+  return (
+    <AuthGate>
+      <>
+        <SettingsPageLayout>
+          {data?.currentUser && !loading && (
+            <div className="forms-container">
+              <SubscriptionForm user={data.currentUser} onSuccess={refetchUser} />
+            </div>
+          )}
+        </SettingsPageLayout>
+        <style jsx>{`
+          .forms-container {
+            width: 100%;
+            max-width: 1008px;
+          }
+          h1 {
+            ${theme.typography.headingLG}
+            text-align: center;
+            margin-bottom: 20px;
+          }
+        `}</style>
+      </>
+    </AuthGate>
+  )
+}
+
+Subscription.getInitialProps = async () => ({
+  namespacesRequired: ['common', 'settings'],
+})
+
+export default withApollo(Subscription)
